+++ agent/src/main/java/com/teamscale/jacoco/agent/testimpact/TestwiseCoverageAgent.java
@@ -6,10 +6,15 @@
 package com.teamscale.jacoco.agent.testimpact;
 
 import com.teamscale.jacoco.agent.AgentBase;
+import com.teamscale.jacoco.agent.GenericExceptionMapper;
 import com.teamscale.jacoco.agent.options.AgentOptions;
+import com.teamscale.jacoco.agent.util.LoggingUtils;
 import com.teamscale.report.testwise.jacoco.JaCoCoTestwiseReportGenerator;
 import org.glassfish.jersey.server.ResourceConfig;
 import org.glassfish.jersey.server.ServerProperties;
+import org.slf4j.Logger;
+
+import java.io.IOException;
 
 /**
  * A wrapper around the JaCoCo Java agent that starts a HTTP server and listens for test events.
@@ -20,6 +25,17 @@ public class TestwiseCoverageAgent extends AgentBase {
 	 * The test event strategy handler.
 	 */
 	protected final TestEventHandlerStrategyBase testEventHandler;
+
+	/** Creates a {@link TestwiseCoverageAgent} based on the given options. */
+	public static TestwiseCoverageAgent create(AgentOptions agentOptions) throws IOException {
+		Logger logger = LoggingUtils.getLogger(JaCoCoTestwiseReportGenerator.class);
+		JaCoCoTestwiseReportGenerator reportGenerator = new JaCoCoTestwiseReportGenerator(
+				agentOptions.getClassDirectoriesOrZips(), agentOptions.getLocationIncludeFilter(),
+				agentOptions.getDuplicateClassFileBehavior(), LoggingUtils.wrap(logger));
+		return new TestwiseCoverageAgent(agentOptions,
+				new TestExecutionWriter(agentOptions.createNewFileInOutputDirectory("test-execution", "json")),
+				reportGenerator);
+	}
 
 
 	public TestwiseCoverageAgent(AgentOptions options, TestExecutionWriter testExecutionWriter,
@@ -39,7 +55,6 @@ public class TestwiseCoverageAgent extends AgentBase {
 				testEventHandler = new CoverageToExecFileStrategy(controller, options, testExecutionWriter);
 				break;
 		}
-
 	}
 
 	@Override
@@ -47,6 +62,6 @@ public class TestwiseCoverageAgent extends AgentBase {
 		ResourceConfig resourceConfig = new ResourceConfig();
 		resourceConfig.property(ServerProperties.WADL_FEATURE_DISABLE, Boolean.TRUE.toString());
 		TestwiseCoverageResource.setAgent(this);
-		return resourceConfig.register(TestwiseCoverageResource.class);
+		return resourceConfig.register(TestwiseCoverageResource.class).register(GenericExceptionMapper.class);
 	}
 }