+++ agent/src/main/java/com/teamscale/jacoco/agent/upload/IUploader.java
@@ -6,8 +6,9 @@ import com.teamscale.report.jacoco.CoverageFile;
 public interface IUploader {
 
 	/**
-	 * Uploads the given coverage file. If the upload was successful, the coverage file on disk will be deleted.
-	 * Otherwise the file is left on disk and a warning is logged.
+	 * Uploads the given coverage file. If the upload was successful, the coverage
+	 * file on disk will be deleted. Otherwise the file is left on disk and a
+	 * warning is logged.
 	 */
 	void upload(CoverageFile coverageFile);