+++ agent/src/main/java/com/teamscale/jacoco/agent/upload/teamscale/TeamscaleUploader.java
@@ -1,20 +1,41 @@
 package com.teamscale.jacoco.agent.upload.teamscale;
 
+import static com.teamscale.jacoco.agent.upload.teamscale.ETeamscaleServerProperties.COMMIT;
+import static com.teamscale.jacoco.agent.upload.teamscale.ETeamscaleServerProperties.MESSAGE;
+import static com.teamscale.jacoco.agent.upload.teamscale.ETeamscaleServerProperties.PARTITION;
+import static com.teamscale.jacoco.agent.upload.teamscale.ETeamscaleServerProperties.PROJECT;
+import static com.teamscale.jacoco.agent.upload.teamscale.ETeamscaleServerProperties.REVISION;
+
+import java.io.File;
+import java.io.IOException;
+import java.io.OutputStreamWriter;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.util.Properties;
+
+import org.conqat.lib.commons.filesystem.FileSystemUtils;
+import org.slf4j.Logger;
+
+import com.google.common.base.Strings;
+import com.teamscale.client.CommitDescriptor;
 import com.teamscale.client.EReportFormat;
 import com.teamscale.client.HttpUtils;
 import com.teamscale.client.ITeamscaleService;
 import com.teamscale.client.TeamscaleServer;
 import com.teamscale.client.TeamscaleServiceGenerator;
+import com.teamscale.jacoco.agent.upload.IUploadRetry;
 import com.teamscale.jacoco.agent.upload.IUploader;
 import com.teamscale.jacoco.agent.util.Benchmark;
 import com.teamscale.jacoco.agent.util.LoggingUtils;
 import com.teamscale.report.jacoco.CoverageFile;
-import org.slf4j.Logger;
-
-import java.io.IOException;
 
 /** Uploads XML Coverage to a Teamscale instance. */
-public class TeamscaleUploader implements IUploader {
+public class TeamscaleUploader implements IUploader, IUploadRetry {
+
+	/**
+	 * The properties file suffix for unsuccessful coverage uploads.
+	 */
+	public static final String RETRY_UPLOAD_FILE_SUFFIX = "_upload-retry.properties";
 
 	/** The logger. */
 	private final Logger logger = LoggingUtils.getLogger(this);
@@ -29,51 +50,88 @@ public class TeamscaleUploader implements IUploader {
 
 	@Override
 	public void upload(CoverageFile coverageFile) {
+		doUpload(coverageFile, this.teamscaleServer);
+	}
+
+	@Override
+	public void reupload(CoverageFile coverageFile, Properties reuploadProperties) {
+		TeamscaleServer server = new TeamscaleServer();
+		server.project = reuploadProperties.getProperty(PROJECT.name());
+		server.commit = CommitDescriptor.parse(reuploadProperties.getProperty(COMMIT.name()));
+		server.partition = reuploadProperties.getProperty(PARTITION.name());
+		server.revision = Strings.emptyToNull(reuploadProperties.getProperty(REVISION.name()));
+		server.userAccessToken = teamscaleServer.userAccessToken;
+		server.userName = teamscaleServer.userName;
+		server.url = teamscaleServer.url;
+		server.setMessage(reuploadProperties.getProperty(MESSAGE.name()));
+		doUpload(coverageFile, server);
+	}
+
+	private void doUpload(CoverageFile coverageFile, TeamscaleServer teamscaleServer) {
 		try (Benchmark benchmark = new Benchmark("Uploading report to Teamscale")) {
-			if (tryUploading(coverageFile)) {
+			if (tryUploading(coverageFile, teamscaleServer)) {
 				deleteCoverageFile(coverageFile);
 			} else {
-				logger.warn("Failed to upload coverage to Teamscale. " +
-								"Won't delete local file {} so you can upload it yourself manually.",
-						coverageFile);
+				logger.warn("Failed to upload coverage to Teamscale. "
+						+ "Won't delete local file {} so that the upload can automatically be retried upon profiler restart. "
+						+ "Upload can also be retried manually.", coverageFile);
+				markFileForUploadRetry(coverageFile);
 			}
 		}
+	}
+
+	@Override
+	public void markFileForUploadRetry(CoverageFile coverageFile) {
+		File uploadMetadataFile = new File(FileSystemUtils.replaceFilePathFilenameWith(
+				com.teamscale.client.FileSystemUtils.normalizeSeparators(coverageFile.toString()),
+				coverageFile.getName() + RETRY_UPLOAD_FILE_SUFFIX));
+		Properties serverProperties = this.createServerProperties();
+		try (OutputStreamWriter writer = new OutputStreamWriter(Files.newOutputStream(uploadMetadataFile.toPath()),
+				StandardCharsets.UTF_8)) {
+			serverProperties.store(writer, null);
+		} catch (IOException e) {
+			logger.warn(
+					"Failed to create metadata file for automatic upload retry of {}. Please manually retry the coverage upload to Teamscale.",
+					coverageFile);
+			uploadMetadataFile.delete();
+		}
+	}
+
+	/**
+	 * Creates server properties to be written in a properties file.
+	 */
+	private Properties createServerProperties() {
+		Properties serverProperties = new Properties();
+		serverProperties.setProperty(PROJECT.name(), teamscaleServer.project);
+		serverProperties.setProperty(PARTITION.name(), teamscaleServer.partition);
+		serverProperties.setProperty(COMMIT.name(), teamscaleServer.commit.toString());
+		serverProperties.setProperty(REVISION.name(), Strings.nullToEmpty(teamscaleServer.revision));
+		serverProperties.setProperty(MESSAGE.name(), teamscaleServer.getMessage());
+		return serverProperties;
 	}
 
 	private void deleteCoverageFile(CoverageFile coverageFile) {
 		try {
 			coverageFile.delete();
 		} catch (IOException e) {
-			logger.warn(
-					"The upload to Teamscale was successful, but the file {} will be left on disk. " +
-							"You can delete it yourself anytime - it is no longer needed.",
-					coverageFile, e);
+			logger.warn("The upload to Teamscale was successful, but the deletion of the coverage file {} failed. "
+					+ "You can delete it yourself anytime - it is no longer needed.", coverageFile, e);
 		}
 	}
 
 	/** Performs the upload and returns <code>true</code> if successful. */
-	private boolean tryUploading(CoverageFile coverageFile) {
+	private boolean tryUploading(CoverageFile coverageFile, TeamscaleServer teamscaleServer) {
 		logger.debug("Uploading JaCoCo artifact to {}", teamscaleServer);
 
 		try {
-			// Cannot be executed in the constructor as this causes issues in WildFly server (See #100)
-			ITeamscaleService api = TeamscaleServiceGenerator.createService(
-					ITeamscaleService.class,
-					teamscaleServer.url,
-					teamscaleServer.userName,
-					teamscaleServer.userAccessToken,
-					HttpUtils.DEFAULT_READ_TIMEOUT,
-					HttpUtils.DEFAULT_WRITE_TIMEOUT
-			);
-			api.uploadReport(
-					teamscaleServer.project,
-					teamscaleServer.commit,
-					teamscaleServer.revision,
-					teamscaleServer.partition,
-					EReportFormat.JACOCO,
-					teamscaleServer.getMessage(),
-					coverageFile.createFormRequestBody()
-			);
+			// Cannot be executed in the constructor as this causes issues in WildFly server
+			// (See #100)
+			ITeamscaleService api = TeamscaleServiceGenerator.createService(ITeamscaleService.class,
+					teamscaleServer.url, teamscaleServer.userName, teamscaleServer.userAccessToken,
+					HttpUtils.DEFAULT_READ_TIMEOUT, HttpUtils.DEFAULT_WRITE_TIMEOUT);
+			api.uploadReport(teamscaleServer.project, teamscaleServer.commit, teamscaleServer.revision,
+					teamscaleServer.partition, EReportFormat.JACOCO, teamscaleServer.getMessage(),
+					coverageFile.createFormRequestBody());
 			return true;
 		} catch (IOException e) {
 			logger.error("Failed to upload coverage to {}", teamscaleServer, e);