+++ impacted-test-engine/src/main/java/com/teamscale/test_impacted/test_descriptor/JUnitVintageTestDescriptorResolver.java
@@ -6,7 +6,7 @@ import org.junit.platform.engine.TestEngine;
 import java.util.Optional;
 
 /** Test default test descriptor resolver for the JUnit vintage {@link TestEngine}. */
-public class JUnitVintageTestDescriptorResolver extends JUnitTestDescriptorResolverBase {
+public class JUnitVintageTestDescriptorResolver extends JUnitClassBasedTestDescriptorResolverBase {
 
 	/** The segment type name that the vintage engine uses for the class descriptor nodes. */
 	public static final String RUNNER_SEGMENT_TYPE = "runner";