+++ system-tests/tia-maven/maven-dump-local-project/src/test/java/foo/IntegIT.java
@@ -1,17 +1,22 @@
 package bar;
 
 import org.junit.jupiter.api.Test;
+
 import foo.SUT;
 
 public class IntegIT {
 
-    @Test
-    public void itBla() throws Exception {
-        new SUT().bla();
-    }
+	@Test
+	public void itBla() throws Exception {
+		new SUT().bla();
+		// Add sleep so we can simulate test execution and test parallel test execution
+		Thread.sleep(2000);
+	}
 
-    @Test
-    public void itFoo() {
-        new SUT().foo();
-    }
+	@Test
+	public void itFoo() throws Exception {
+		new SUT().foo();
+		// Add sleep so we can simulate test execution and test parallel test execution
+		Thread.sleep(2000);
+	}
 }